src/requirements.txt
Flask
urllib3
requests_oauthlib
python-twitter
flask-sqlalchemy
Flask-Session
